+++ fixtures/architect/src/http/any-catchall/index.js
@@ -1,2 +1,2 @@
 const { createRequestHandler } = require("@remix-run/architect");
-exports.handler = createRequestHandler({ enableSessions: true });
+exports.handler = createRequestHandler();