+++ fixtures/gists-app/server.js
@@ -1,6 +1,5 @@
 const express = require("express");
 const morgan = require("morgan");
-const session = require("express-session");
 const { createRequestHandler } = require("@remix-run/express");
 
 const port = process.env.PORT || 3000;
@@ -12,8 +11,6 @@ app.use(express.static("public"));
 if (process.env.NODE_ENV !== "production" && process.env.NODE_ENV !== "test") {
   app.use(morgan("dev"));
 }
-
-app.use(session({ secret: "remix", resave: false, saveUninitialized: false }));
 
 // server-side redirect
 app.get("/user-gists/:username", (req, res) => {