+++ fixtures/tutorial/server.js
@@ -1,6 +1,5 @@
 const express = require("express");
 const morgan = require("morgan");
-// const session = require("express-session");
 const { createRequestHandler } = require("@remix-run/express");
 
 let app = express();
@@ -10,18 +9,10 @@ if (process.env.NODE_ENV === "development") {
 }
 
 app.use(express.static("public"));
-// app.use(
-//   session({
-//     secret: "remix",
-//     resave: false,
-//     saveUninitialized: false,
-//   })
-// );
 
 app.all(
   "*",
   createRequestHandler({
-    enableSessions: false,
     getLoadContext() {
       // Whatever you return here will be passed as `context` to your loaders.
     }