+++ packages/architect/index.ts
@@ -1,11 +1,6 @@
-import arc from "@architect/functions";
 import { URL } from "url";
-import {
-  createAdapter,
-  Request,
-  createSession,
-  createSessionFacade
-} from "@remix-run/core";
+import { createAdapter, Request } from "@remix-run/core";
+
 import "./fetchGlobals";
 
 import type {
@@ -29,40 +24,11 @@ export let createRequestHandler = createAdapter({
     });
   },
 
-  async sendPlatformResponse(
-    remixResponse,
-    remixSession
-  ): Promise<ArcResponse> {
-    // Don't write a cookie if the session object is empty. It always has
-    // `iam` on it (that's why it's length > 1), need to look into this more,
-    // docs don't talk about it.
-    if (Object.keys(remixSession.mutableData).length > 1) {
-      remixResponse.headers.append(
-        "set-cookie",
-        await arc.http.session.write(remixSession.mutableData)
-      );
-    }
-
+  async sendPlatformResponse(remixResponse): Promise<ArcResponse> {
     return {
       statusCode: remixResponse.status,
       headers: Object.fromEntries(remixResponse.headers),
       body: await remixResponse.text()
     };
-  },
-
-  async createRemixSession(enableSessions: boolean, req: ArcRequest) {
-    if (enableSessions === false) {
-      return createSessionFacade(
-        "You need to enable sessions to use them. Enable with `createRequestHandler({ enableSessions: true })` in your Remix arc handler definition."
-      );
-    }
-
-    let arcSession = await arc.http.session.read(req);
-    return createSession(arcSession, () => {
-      // TODO: Make sure this works for a `logout` situation
-      // Delete everything but `iam`, which they seem to keep around always, not
-      // a lot of documentation here
-      arcSession = { iam: arcSession.iam };
-    });
   }
 });