+++ packages/express/__tests__/index-test.ts
@@ -1,14 +1,6 @@
 import express from "express";
 import supertest from "supertest";
 
-// @ts-ignore Adding @types/express-session breaks the build of
-// @remix-run/core.assetServer, because `app.get("*", handleRequest)` wants a
-// request type with session, but assetServer imports the non-augmented types,
-// and since our node_modules are shared at the root of the project, we either
-// have to tell the asset server it's got a req.session, or just ignore this and
-// not install @types/express-session so it doesn't do weird things to `app.get`.
-import session from "express-session";
-
 import { createRequestHandler } from "../index";
 
 import { Response } from "@remix-run/core/fetch";
@@ -36,7 +28,7 @@ describe("express createRequestHandler", () => {
     beforeEach(() => {
       mockedReadConfig.mockResolvedValue({} as ReturnType<typeof readConfig>);
       let app = express();
-      app.all("*", createRequestHandler({ enableSessions: false }));
+      app.all("*", createRequestHandler());
       request = supertest(app);
     });
 
@@ -78,62 +70,6 @@ describe("express createRequestHandler", () => {
 
       let res = await request.get("/");
       expect(res.headers["x-time-of-year"]).toBe("most wonderful");
-    });
-  });
-
-  describe("sessions", () => {
-    it("sets and gets values from a cookie session", async () => {
-      mockedReadConfig.mockResolvedValue({} as ReturnType<typeof readConfig>);
-      mockedCreateRequestHandler.mockImplementation(
-        () => async (req, session) => {
-          if (req.url.endsWith("/set")) {
-            session.set("foo", "🍪");
-            return new Response("Set");
-          } else if (req.url.endsWith("/get")) {
-            return new Response(`Session: ${session.get("foo")}`);
-          } else if (req.url.endsWith("/destroy")) {
-            await session.destroy();
-            return new Response(`Session destroyed`);
-          } else {
-            throw new Error("unknown test url");
-          }
-        }
-      );
-
-      let cookie = "";
-      let app = express();
-      app.use(
-        session({
-          secret: "remix",
-          resave: false,
-          saveUninitialized: false
-        })
-      );
-      app.all("*", createRequestHandler());
-      let request = supertest(app);
-
-      let setRes = await request.get("/set");
-      expect(setRes.headers["set-cookie"]).toBeDefined();
-      cookie = setRes.headers["set-cookie"].pop().split(";")[0];
-
-      let getRes = await request.get("/get").set("Cookie", cookie);
-      expect(getRes.text).toBe("Session: 🍪");
-
-      // TODO: Test destroying
-      // For some reason the "set-cookie" header doesn't come back on the
-      // `getRes`. I think it's something to do with supertest making a new
-      // server for pretty much every request, might need to figure out how to
-      // get it to be the same server across requests. I know it works over in
-      // the fixtures, so just marking as a todo for now. Already wasted a
-      // couple hours here, so I'm moving on. -ryan
-
-      // This shouldn't be empty but it is!
-      // console.log({ getResCookie: getRes.headers["set-cookie"] })
-
-      // This will pass, but that's only because every request after the first
-      // doesn't set a cookie anyway :(
-      // let destroyRes = await request.get("/destroy").set("Cookie", cookie);
-      // expect(destroyRes.headers["set-cookie"]).toBeUndefined();
     });
   });
 });